moves module injector to main injector
